Adds all the shift_manager module methods to decryptor and makes adjust index method for test
