+++ lib/shift_manager.rb
@@ -36,4 +36,48 @@ module ShiftManager
     extracted_digits = extracted_digits_backwards.reverse
     extracted_digits.join
   end
+
+  def shift_a(message_characters, shifts)
+    message_characters.map.with_index do |char, ind|
+      if ind % 4 == 0
+        new_index = adjust_index(char, shifts[:A])
+        char = @characters[new_index]
+      else
+        char
+      end
+    end.join
+  end
+
+  def shift_b(message_characters, shifts)
+    message_characters.map.with_index do |char, ind|
+      if (ind - 1) % 4 == 0
+        new_index = adjust_index(char, shifts[:B])
+        char = @characters[new_index]
+      else
+        char
+      end
+    end.join
+  end
+
+  def shift_c(message_characters, shifts)
+    message_characters.map.with_index do |char, ind|
+      if (ind - 2) % 4 == 0
+        new_index = adjust_index(char, shifts[:C])
+        char = @characters[new_index]
+      else
+        char
+      end
+    end.join
+  end
+
+  def shift_d(message_characters, shifts)
+    message_characters.map.with_index do |char, ind|
+      if (ind - 3) % 4 == 0
+        new_index = adjust_index(char, shifts[:D])
+        char = @characters[new_index]
+      else
+        char
+      end
+    end.join
+  end
 end
